Make asteroidresponder succulent
The response structure is now

PHP Request -> Python writing something to file -> C++ reading from file, searching through an already-loaded set of data, and writing what it finds to file -> Python detects this and gives it to PHP

Python is called with -S in an attempt to minimize startup overhead. It seems the average search time is 14-17 ms, most of which is spent in native level, which is quite acceptable.